moved enums to common one
