+++ AlumniManagementPortal/src/main/java/com/miu/alumnimanagementportal/entities/News.java
@@ -1,9 +1,8 @@
 package com.miu.alumnimanagementportal.entities;
 
+import com.miu.alumnimanagementportal.common.enums.NewsType;
 import jakarta.persistence.Entity;
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
 
 @Entity
 @Data
